+++ Changes.md
@@ -6,6 +6,7 @@ Version 402.1.1 - TBR
 1. Fix 'section_activity_summary' method.
 2. Single section navigation fixes.
 3. Improve single section navigation appearance.
+4. Add bulk editing tools.
 
 Version 402.1.0 - 07/05/23
 --------------------------